requirements.txt
asgiref==3.8.1
Django==5.1.6
django-cors-headers==4.7.0
django-extensions==3.2.3
djangorestframework==3.16.0
graphviz==0.20.3
gunicorn==23.0.0
mysqlclient==2.2.7
packaging==24.2
sqlparse==0.5.3
tzdata==2025.1
waitress==3.0.2

gunicorn
django
djangorestframework
axios